requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic
azure-storage-blob
openai
pymupdf
python-docx
pandas
python-dotenv
